fix: preserve input dtype in apply_eq_mono and apply_eq_gains (issue #2200)
apply_eq_mono() always cast its output to np.float32 regardless of the
input dtype, silently truncating float64 audio and introducing quantization
noise in high-precision processing chains.

Two-line fix in filters.py:
1. apply_eq_mono: dtype=np.float32 → dtype=audio_mono.dtype on the return cast
2. apply_eq_gains: np.zeros() padding path now passes dtype=audio_chunk.dtype so
   float32 input is not promoted to float64 before reaching apply_eq_mono, which
   would have caused the corrected return cast to yield float64 for float32 input.

Seven new tests in TestApplyEQGainsDtypePreservation cover mono/stereo,
normal and padded (short-chunk) paths for both float32 and float64 inputs,
plus a flat-gain round-trip precision check for float64.

+++ auralis/dsp/eq/filters.py
@@ -35,8 +35,9 @@ def apply_eq_gains(audio_chunk: np.ndarray,
     original_len = len(audio_chunk)
 
     if original_len < fft_size:
-        # Pad with zeros for processing
-        padded = np.zeros((fft_size, audio_chunk.shape[1] if audio_chunk.ndim == 2 else 1))
+        # Pad with zeros for processing; preserve dtype so float32 isn't promoted.
+        padded = np.zeros((fft_size, audio_chunk.shape[1] if audio_chunk.ndim == 2 else 1),
+                          dtype=audio_chunk.dtype)
         if audio_chunk.ndim == 2:
             padded[:original_len, :] = audio_chunk
         else:
@@ -98,7 +99,7 @@ def apply_eq_mono(audio_mono: np.ndarray,
     # Transform back to time domain
     processed_audio = np.real(ifft(spectrum))
 
-    return np.asarray(processed_audio[:len(audio_mono)], dtype=np.float32)
+    return np.asarray(processed_audio[:len(audio_mono)], dtype=audio_mono.dtype)
 
 
 def create_filter_bank(critical_bands: list[Any],
